requirements.txt
# MCP Inbox — dependencies
fastmcp>=2.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
google-api-python-client>=2.126.0
google-auth-oauthlib>=1.2.0
slack-sdk>=3.27.0
python-telegram-bot[socks]>=21.3
pydantic-settings>=2.2.1
aiosqlite>=0.20.0
python-dotenv>=1.0.1
jinja2>=3.1.4
websockets>=12.0
httpx>=0.27.0
telethon>=1.36.0
PySocks>=1.7.1
